+++ R/check.R
@@ -21,15 +21,21 @@ check_class <- function(x, what) {
 check_cols <- function(cols, data_names) {
   cols_obs <- cols %in% data_names
   cols_mis <- cols[!cols_obs]
-  stopifnot_(
-    all(cols_obs),
-    c(
-      "The columns {.val {cols}} must exist in the data.",
-      `x` = "The following columns were
-             not found in the data: {.val {cols_mis}}."
-    )
-  )
+  if (length(cols_mis) > 0) {
+    if (length(cols) == 1) {
+      stop_("The column {.val {cols_mis}} was not found in the data.")
+    } else {
+      stop_(
+        c(
+          "The columns {.val {cols}} must exist in the data.",
+          `x` = "The following columns were
+          not found in the data: {.val {cols_mis}}."
+        )
+      )
+    }
+  }
 }
+
 
 #' Check That `x` is a Logical Value
 #'