+++ NEWS.md
@@ -1,6 +1,8 @@
 # codyna 0.2.0
 
-* Added lift measure to `discover_patterns()`.
+* Added a lift measure to `discover_patterns()`.
+* The `min_count` argument of `discover_patterns()` is now called `min_freq`. In the context of this function, "count" now refers to the number of sequences that a pattern is present in, and "frequency" is the total number of occurrences of a pattern.
+* Added the argument `group` to `discover_patterns()` for comparing pattern counts between groups.
 
 # codyna 0.1.0